+++ src/test/java/com/mageddo/dnsproxyserver/server/dns/solver/SolversCacheTest.java
@@ -9,6 +9,7 @@ import org.xbill.DNS.Flags;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @ExtendWith(MockitoExtension.class)
@@ -33,6 +34,19 @@ class SolversCacheTest {
     assertEquals(req.getHeader().getID(), res.getHeader().getID());
     assertTrue(header.getFlag(Flags.QR));
 
+  }
+
+  @Test
+  void cantCacheWhenDelegateSolverHasNoAnswer(){
+    // arrange
+    final var query = MessageTemplates.acmeAQuery();
+
+    // act
+    final var res = this.cache.handle(query, message -> null);
+
+    // assert
+    assertNull(res);
+    assertEquals(0, this.cache.getSize());
   }
 
 }