+++ src/test/java/com/mageddo/dnsproxyserver/templates/ConfigTemplates.java
@@ -1,11 +1,17 @@
 package com.mageddo.dnsproxyserver.templates;
 
 import com.mageddo.dnsproxyserver.config.Config;
+import com.mageddo.dnsproxyserver.config.entrypoint.LogLevel;
 
 import java.nio.file.Paths;
 
 public class ConfigTemplates {
   public static Config withoutId() {
+    return defaultBuilder()
+      .build();
+  }
+
+  private static Config.ConfigBuilder defaultBuilder() {
     return Config
       .builder()
       .logFile("/tmp/dps.log")
@@ -19,6 +25,9 @@ public class ConfigTemplates {
       .version("3.0.0")
       .dnsServerPort(53)
       .domain("com")
-      .build();
+      .logLevel(LogLevel.WARNING)
+      .resolvConfPath(Paths.get("/etc/resolv.conf"));
   }
+
+
 }